Добавила реализацию PostService через интерфейс.
